Add files via upload
Python script to measure the execution time of a naive Traveling Salesman Problem (TSP) algorithm across different numbers of cities, ranging from 2 to 9. To simplify the graph creation, I used a uniform distance of 10 units between any two different cities and considered the distance to be infinite, for the same city to avoid self-loops. I approximated the TSP using uniform distances to focus on the time complexity rather than accurate pathfinding.

The resulting plot shows a sharp increase in time spent as the number of cities increases, particularly between 7 and 9 cities. This illustrates the factorial time complexity of the TSP algorithm.